saas/api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0
python-multipart>=0.0.6
httpx>=0.26.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.115.0
